+++ .github/workflows/test and deploy.yml
@@ -14,30 +14,19 @@ jobs:
       - run: npm run build
       - run: npm run format:check
       - run: npm run lint:check
-  gcloud:
-    needs: check
-    name: gcloud deploy
+  build-and-push:
     runs-on: ubuntu-latest
+    needs:
+      - check
     steps:
-      - uses: actions-hub/gcloud@master
-        env:
-          PROJECT_ID: ${{ secrets.GCLOUD_PROJECT_ID }}
-          APPLICATION_CREDENTIALS: ${{ secrets.GCLOUD_APPLICATION_CREDENTIALS }}
+      - uses: actions/checkout@v2
+      - uses: actions/setup-node@v1
         with:
-          args: functions deploy gumroadalerts --entrypoint
-  deploy:
-    if: github.ref == 'refs/heads/master'
-    name: deploy
-    needs: check
-    runs-on: ubuntu-latest
-    steps:
-      - name: Create SSH key
-        run: |
-          mkdir -p ~/.ssh/
-          cd ~/.ssh/
-          echo "${{secrets.SSH_PRIVATE_KEY}}" > ./id_rsa
-          sudo chmod 600 ./id_rsa
-        shell: bash
-      - name: deploy
-        run: |
-          ssh ${{ secrets.SSH_USER }}@${{ secrets.SSH_HOST }} -o StrictHostKeyChecking=no "cd /home/gumroadalerts && git pull && npm run setup"
+          node-version: 15.4
+      - run: npm i serverless@2.39.0
+      - run: npx serverless deploy
+    env:
+      AWS_ACCOUNT_ID: ${{ secrets.AWS_ACCOUNT_ID }}
+      AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
+      AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+      APP_CONFIG: ${{ secrets.APP_CONFIG }}
