+++ .idea/jsonSchemas.xml
@@ -55,6 +55,21 @@
             </SchemaInfo>
           </value>
         </entry>
+        <entry key="serverless">
+          <value>
+            <SchemaInfo>
+              <option name="name" value="serverless" />
+              <option name="relativePathToSchema" value="https://raw.githubusercontent.com/softprops/serverless-yml-schema/master/serverless-schema.json" />
+              <option name="patterns">
+                <list>
+                  <Item>
+                    <option name="path" value="serverless.yml" />
+                  </Item>
+                </list>
+              </option>
+            </SchemaInfo>
+          </value>
+        </entry>
       </map>
     </state>
   </component>
