+++ .prettierignore
@@ -2,5 +2,5 @@ node_modules
 dist
 package.json
 package-lock.json
-*_files.json
 **/lcdev__app-config/*
+.serverless
