+++ serverless.yml
@@ -1,0 +1,93 @@
+# Welcome to Serverless!
+#
+# This file is the main config file for your service.
+# It's very minimal at this point and uses default values.
+# You can always add more config options for more control.
+# We've included some commented out config examples here.
+# Just uncomment any of them to get that config option.
+#
+# For full config options, check the docs:
+#    docs.serverless.com
+#
+# Happy Coding!
+
+service: gumroadalerts
+# app and org for use with dashboard.serverless.com
+#app: your-app-name
+#org: your-org-name
+
+# You can pin your service to only deploy with a specific Serverless version
+# Check out our docs for more details
+frameworkVersion: '2.39.0'
+
+provider:
+  name: aws
+  timeout: 120
+  ecr:
+    images: # Definitions of images that later can be referenced by key in `function.image`
+      image:
+        path: ./
+#  # you can overwrite defaults here
+#  #  stage: dev
+#  #  region: us-east-1
+#  # you can add statements to the Lambda function's IAM Role here
+#  iamRoleStatements:
+#    #    - Effect: "Allow"
+#    #      Action:
+#    #        - "s3:ListBucket"
+#    #      Resource: { "Fn::Join" : ["", ["arn:aws:s3:::", { "Ref" : "ServerlessDeploymentBucket" } ] ]  }
+#    - Effect: 'Allow'
+#      Action:
+#        - 's3:PutObject'
+#        - 's3:DeleteObject'
+#      Resource:
+#        Fn::Join:
+#          - ''
+#          - - 'arn:aws:s3:::'
+#            - 'Ref': 'gumroadalerts'
+#            - '/*'
+
+# you can define service wide environment variables here
+#  environment:
+#    variable1: value1
+
+# you can add packaging information here
+#package:
+#  patterns:
+#    - '!exclude-me.js'
+#    - '!exclude-me-dir/**'
+#    - include-me.js
+#    - include-me-dir/**
+
+functions:
+  gumroadalerts:
+    image: image
+    events:
+      - schedule: rate(15 minutes)
+    environment:
+      APP_CONFIG: ${env:APP_CONFIG}
+
+resources:
+  Resources:
+    S3Bucket:
+      Type: AWS::S3::Bucket
+      Properties:
+        BucketName: gumroadalerts
+    BucketPolicy:
+      Type: AWS::S3::BucketPolicy
+      Properties:
+        PolicyDocument:
+          Statement:
+            - Sid: PublicReadForGetBucketObjects
+              Effect: Allow
+              Principal:
+                AWS: 'arn:aws:iam::${env:AWS_ACCOUNT_ID}:role/gumroadalerts-dev-us-east-1-lambdaRole'
+              Action:
+                - 's3:GetObject'
+                - 's3:PutObject'
+                - 's3:DeleteObject'
+                - 's3:GetObjectAcl'
+                - 's3:PutObjectAcl'
+              Resource: arn:aws:s3:::gumroadalerts/*
+        Bucket:
+          Ref: S3Bucket
